# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deokhugam API",
    "description": "Deokhugam API 문서",
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/users": {
      "post": {
        "tags": [
          "사용자 관리"
        ],
        "summary": "회원가입",
        "description": "새로운 사용자를 등록합니다.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "회원가입 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "409": {
            "description": "이메일 중복",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (입력값 검증 실패)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{userId}": {
      "get": {
        "tags": [
          "사용자 관리"
        ],
        "summary": "사용자 상세 정보 조회",
        "description": "사용자 ID로 상세 정보를 조회합니다.",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "사용자 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "사용자 정보 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "사용자 정보 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "CustomHeaderAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "사용자 관리"
        ],
        "summary": "사용자 논리 삭제",
        "description": "본인이 작성한 사용자을 논리적으로 삭제합니다.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "사용자 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "사용자 삭제 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "사용자 정보 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403": {
            "description": "사용자 삭제 권한 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "CustomHeaderAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "사용자 관리"
        ],
        "summary": "사용자 정보 수정",
        "description": "사용자의 닉네임을 수정합니다.",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "사용자 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "사용자 정보 수정 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (입력값 검증 실패)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "사용자 정보 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403": {
            "description": "사용자 정보 수정 권한 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "CustomHeaderAuth": []
          }
        ]
      }
    },
...
    "/api/users/power": {
      "get": {
        "tags": [
          "사용자 관리"
        ],
        "summary": "인기 사용자 목록 조회",
        "description": "기간별 인기 사용자 목록을 조회합니다",
        "operationId": "getPopularUsers",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "description": "랭킹 기간",
            "required": false,
            "schema": {
              "type": "string",
              "default": "DAILY",
              "enum": [
                "DAILY",
                "WEEKLY",
                "MONTHLY",
                "ALL_TIME"
              ]
            }
          },
          {
            "name": "direction",
            "in": "query",
            "description": "정렬 방향",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ASC"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "커서 페이지네이션 커서",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "description": "보조 커서(createdAt)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "페이지 크기",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "인기 사용자 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PopularReviewDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (랭킹 기간 오류, 정렬 방향 오류)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PopularReviewDto"
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PopularReviewDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/books/{bookId}/hard": {
      "delete": {
        "tags": [
          "도서 관리"
        ],
        "summary": "도서 물리 삭제",
        "description": "도서를 물리적으로 삭제합니다.",
        "operationId": "hardDelete",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "도서 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "9a4d6e2b-8f53-4c8c-a7c7-21d5079b8b37"
          }
        ],
        "responses": {
          "204": {
            "description": "도서 삭제 성공"
          },
          "404": {
            "description": "도서 정보 없음"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserCreateRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "nickname": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "pattern": "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[@$!%*#?&])[A-Za-z\\d@$!%*#?&]{8,100}$"
          }
        },
        "required": [
          "email",
          "nickname",
          "password"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nickname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean"
          }
        }
      },
...
      "UserUpdateRequest": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "maxLength": 20,
            "minLength": 2
          }
        },
        "required": [
          "nickname"
        ]
      },
...
      "PopularReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reviewId": {
            "type": "string",
            "format": "uuid"
          },
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "bookTitle": {
            "type": "string"
          },
          "bookThumbnailUrl": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "userNickname": {
            "type": "string"
          },
          "reviewContent": {
            "type": "string"
          },
          "reviewRating": {
            "type": "number",
            "format": "double"
          },
          "period": {
            "type": "string",
            "enum": [
              "DAILY",
              "WEEKLY",
              "MONTHLY",
              "ALL_TIME"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "rank": {
            "type": "integer",
            "format": "int64"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "likeCount": {
            "type": "integer",
            "format": "int64"
          },
          "commentCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "CustomHeaderAuth": {
        "type": "apiKey",
        "name": "Deokhugam-Request-User-ID",
        "in": "header"
      }
    }
  }
}